scripts mejorados

+++ scripts/scriptAreas.ps1
@@ -1,4 +1,4 @@
-$rootDir = "D:/Work/Diseño/diseño"
+$rootDir = "D:/Work/Design/prevengos"
 $sourcePath
 
 $subAreas = @(
@@ -424,40 +424,46 @@ $subAreas = @(
 
 foreach ($subArea in $subAreas) {
     $destinationPath = $subArea.link
-    $name = $subArea.name
-    $title = "Prevengos - " + $name.Substring(0, 1).ToUpper() + $name.Substring(1).ToLower()
 
-    # Si createAnclas es true, agregar el apartado de las anclas al lateral
-    if ($subArea.createAnclas -eq $true) {
-        $sourcePath = "D:/Work/Diseño/diseño/html/templates/template.html"
+    if (-not (Test-Path $destinationPath)) {
+        $name = $subArea.name
+        $title = "Prevengos - " + $name.Substring(0, 1).ToUpper() + $name.Substring(1).ToLower()
+
+        # Si createAnclas es true, agregar el apartado de las anclas al lateral
+        if ($subArea.createAnclas -eq $true) {
+            $sourcePath = $rootDir + "/html/templates/template.html"
+        }
+        else {
+            $sourcePath = $rootDir + "/html/templates/templateWithSidebarHidden.html"
+        }
+
+        # Copiar el archivo original
+        Copy-Item $sourcePath $destinationPath -Force
+
+        # Esperar un momento para asegurar que el archivo esté disponible
+        Start-Sleep -Milliseconds 500
+
+        # Leer el contenido del archivo copiado
+        $htmlContent = Get-Content -Path $destinationPath -Raw
+
+        # Mostrar el contenido antes de la modificación (solo para depuración)
+        # Write-Host "Antes: `$htmlContent"
+
+        # Reemplazar el título en la etiqueta <title>
+        $htmlContent = $htmlContent -replace '<title>.*?</title>', "<title>$title</title>"
+
+        # Reemplazar el texto dentro del <h1> en la sección Page Title
+        $htmlContent = $htmlContent -replace '(<h1[^>]*?>)\s*[^<]+(\s*</h1>)', "`$1$name`$2"
+
+        # Mostrar el contenido después de la modificación (solo para depuración)
+        # Write-Host "Después: `$htmlContent"
+
+        # Guardar el archivo modificado
+        $htmlContent | Set-Content -Path $destinationPath -Encoding UTF8 -Force
+
+        Write-Host "Archivo creado y modificado en: $destinationPath"
     }
     else {
-        $sourcePath = "D:/Work/Diseño/diseño/html/templates/templateWithSidebarHidden.html"
+        Write-Host "El archivo ya existe: $destinationPath"
     }
-
-    # Copiar el archivo original
-    Copy-Item $sourcePath $destinationPath -Force
-
-    # Esperar un momento para asegurar que el archivo esté disponible
-    Start-Sleep -Milliseconds 500
-
-    # Leer el contenido del archivo copiado
-    $htmlContent = Get-Content -Path $destinationPath -Raw
-
-    # Mostrar el contenido antes de la modificación (solo para depuración)
-    # Write-Host "Antes: `$htmlContent"
-
-    # Reemplazar el título en la etiqueta <title>
-    $htmlContent = $htmlContent -replace '<title>.*?</title>', "<title>$title</title>"
-
-    # Reemplazar el texto dentro del <h1> en la sección Page Title
-    $htmlContent = $htmlContent -replace '(<h1[^>]*?>)\s*[^<]+(\s*</h1>)', "`$1$name`$2"
-
-    # Mostrar el contenido después de la modificación (solo para depuración)
-    # Write-Host "Después: `$htmlContent"
-
-    # Guardar el archivo modificado
-    $htmlContent | Set-Content -Path $destinationPath -Encoding UTF8 -Force
-
-    Write-Host "Archivo creado y modificado en: $destinationPath"
 }
